fix makeing new

+++ database/migrations/2025_12_26_184736_add_is_admin_to_users_table.php
@@ -10,9 +10,11 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::table("users", function (Blueprint $table) {
-            $table->boolean("is_admin")->default(false)->after("email");
-        });
+        if (!Schema::hasColumn("users", "is_admin")) {
+            Schema::table("users", function (Blueprint $table) {
+                $table->boolean("is_admin")->default(false)->after("email");
+            });
+        }
     }
 
     /**
@@ -20,8 +22,10 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table("users", function (Blueprint $table) {
-            $table->dropColumn("is_admin");
-        });
+        if (Schema::hasColumn("users", "is_admin")) {
+            Schema::table("users", function (Blueprint $table) {
+                $table->dropColumn("is_admin");
+            });
+        }
     }
 };
